\name{djia}
\alias{djia}
\docType{data}
\title{
Dow Jones Industrial Average
}
\description{
High, low, opening, and closing prices and trading volume
for the Dow Jones Industrial Average.

The data set has:
\itemize{
\item The closing price only from 1915 through September 1928.
\item The high, low, and closing prices from
October 1928 through March 9, 1984.
\item The high, low, opening, and closing prices
from March 12, 1984 through December 1986.
\item The high, low, opening, and closing prices
and the trading volume from January 1987 through February 1990.
}
}
\format{
An object of class \code{timeSeries} with the high, low, open and
  close prices stored as a \code{data.frame}.
}
\source{
Downloaded from the College of Business, Ohio State University
web site in early 1990.
}
\keyword{datasets}
