\name{exch.rate}
\alias{exch.rate}
\title{
  Foreign Exchange Rates
}
\description{
Exchange rates between the US Dollar and the 
\itemize{
\item British Pound (GBP)
\item Canadian Dollar (CAD)
\item German Mark (DEM)
\item Japanese Yen (JPY)
\item Swiss Franc (CHF)
}
in a multi-variate time series.

Data from Andreas S. Weigend, Bernardo A.
Huberman, and David
E. Rumelhart, "Predicting Sunspots and Exchange Rates with
Connectionist Networks", pp. 395-432 in M. Casdagli and S.
Eubank, eds, Nonlinear Modeling and Forecasting, Addison-Wesley,
1992.  

Extracted from the Andreas Weigend web site
http://www.stern.nyu.edu/~aweigend/Time-Series/TSWelcome.html.
}
\keyword{sysdata}
% docclass is function
% Converted by mySd2Rd version 47442.
