\name{fed.rate}
\alias{fed.rate}
\docType{data}
\title{
Federal Reserve Interest Rates
}
\description{
Interest rate data from the web site of the Federal Reserve Bank,
\url{https://www.federalreserve.gov/releases/h15/data.htm},
running from 1972 to 1997.  

You can find more information on that web site. 

Documentation below is derived from the data files.

\code{fed.rate}  is a multi-variate time series, with the following
columns:
}
\value{

\item{prime.rate}{
Bank Prime Loan Rate, daily including weekends and holidays.
}
\item{discount.rate}{
Discount Rate for the Federal Reserve Bank of New York,
daily including weekends and holidays, which is the simple interest rate
at which depository institutions borrow from the Federal Reserve Bank
of New York.
}
\item{fedfunds.rate}{
Federal Funds Effective Rate, daily including weekends and holidays,
which is the cost of borrowing immediately available funds, primarily
for one day. The effective rate is a weighted average of the
reported rates at which different amounts of the
day's trading through New York brokers occurs.
}
\item{mortgage.rate}{
Conventional Mortgage Rates for fixed-rate mortgages from the Federal
Home Loan Mortgage Corporation, weekly on Fridays.
}
}
\keyword{sysdata}
% docclass is data
% Converted by mySd2Rd version 47442.
