\name{net.packet}
\alias{net.packet}
\docType{data}
\title{
  Network Packet Traffic
}
\description{
Time, type, and size for 10,000 network packets
on a local intranet, just before 5PM on March 27, 1998, as
reported by the Unix snoop command. Size is available only
for packets whose type is \code{"TCP"} .
}
\format{
  An object of class \code{timeSeries} with the high, low, open and
  close prices stored as a \code{data.frame}.
  }
\keyword{datasets}
