\name{say.wavelet}
\alias{say.wavelet}
\docType{data}
\title{
  Speech Signal
}
\description{
A signal of a voice saying the word \code{"wavelet"} , sampled at 11025 Hz.
There are 8192 samples, ranging in time from 0 to approximately
0.7429 seconds.  
}
\format{
  The signal is of class \code{signalSeries} .
  }
\keyword{datasets}

