\name{tbond}
\alias{tbond}
\title{
Treasury Bond Futures Trading Data
}
\description{
Treasury Bond futures trading data: high and low prices
over 20-minute intervals from January 7, 1994, to February
3, 1995. 

Taken from the Ohio State University web site
http://www.cob.ohio-state.edu/~fin/osudown.htm
(downloaded downloaded in early 1996).
The dataset is a two-column time series.
It is used to illustrate a drop in bond prices that
occurred in 1994.
}
\keyword{sysdata}
