\name{tbond}
\alias{tbond}
\docType{data}
\title{
Treasury Bond Futures Trading Data
}
\description{
Treasury Bond futures trading data: high and low prices
over 20-minute intervals from January 7, 1994, to February 3, 1995. 

It is used to illustrate a drop in bond prices that occurred in 1994.
}
\format{
An object of class \code{timeSeries} with the high and low prices
stored as a \code{data.frame}.
}
\source{
Downloaded from the Fisher College of Business, Ohio State University
web site in early 1996.
}
\keyword{sysdata}
